more

+++ python/sglang/srt/layers/moe/ep_moe/layer.py
@@ -439,8 +439,8 @@ class DeepEPMoE(EPMoE):
         forward_batch: ForwardBatch,
     ):
         def get_tensor_info(x):
-            min = x.min() if x.numel() > 0 else None
-            max = x.max() if x.numel() > 0 else None
+            min = x.float().min() if x.numel() > 0 else None
+            max = x.float().max() if x.numel() > 0 else None
             mean = x.float().mean() if x.numel() > 0 else None
             return f"shape={x.shape} dtype={x.dtype} device={x.device} stride={x.stride()} min={min} max={max} mean={mean}"
         print(f"[{torch.distributed.get_rank()}, {self.layer_id=}] hi call dispatch {get_tensor_info(hidden_states)=}")
